requirements.txt
Django==3.0.10
django-rest-auth==0.9.5
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
mysqlclient==2.0.1
Pillow==8.0.0
PyJWT==1.7.1
pytz==2020.1
six==1.15.0

